MIQ-2274: Add datasource info for published slices
When replicating dashboards with charts, the subscriber layer lists charts with duplicated names which makes harder for the user to link the publisher and subsciber correctly. While motive of replication of dashboards is to use it with different datasources, so adding the datasource of the slices in the filterbox selector would improve experience of the administrator who's working the dashboards.

This commit adds datasource info appended with the slice's name in the subscriber layer's "Select chart" filter.